refactor(snack): move snack to entities
