+++ src/entities/index.ts
@@ -7,5 +7,5 @@ export { PaymentServiceRow } from './payment-services'
 export { ChipsList } from './chips-list'
 export { AmountInput } from './amout-input'
 export { $paymentCategories, fetchPaymentCategoriesFx } from "./payments"
-export { Snack } from "../app/snack"
-export { $snackMessage, $snackShowing, changeSnackShowing, changeSnackMessage } from "../app/snack"
+export { Snack } from "./snack"
+export { addSnack, $snacks, removeShowedSnack } from "./snack"
