+++ src/entities/snack/index.ts
