+++ src/entities/snack/model/index.ts
