+++ src/entities/snack/model/store.ts
