+++ src/entities/snack/ui/index.ts
