+++ src/entities/snack/ui/snack.tsx
